feat: add an explicit input for the target deployment prefix (#9)
Add an `aws_target_prefix` input variable, which must match the
deployment prefix associated with a target Stacklet instance. This allows the
existing `prefix` variable to keep its intended use - a chunk of text
that will be prepended to resource names.

Modify `local.audience` so that it incorporates `aws_target_prefix`
rather than `prefix`. As far as I can tell the remaining uses of
`prefix` are fine as-is.