Improve function security settings
Enforce HTTPS with TLS 1.3 on the HTTP endpoint, even though the data
plane aspect of it is unused and we'd really like to disable it if Azure
permitted. Zip deployments will use it, at least, since they use the SCM
endpoint which seems to use the same settings.

Also disable the friendly welcome page on the function domain, as it
suggests things aren't set up properly. It'll now return 204 No Content
instead.